@@ -1,24 +1,23 @@
-﻿namespace NetEscapades.EnumGenerators
+﻿namespace NetEscapades.EnumGenerators;
+
+/// <summary>
+/// Add to enums to indicate that extension methods should be generated for the type.
+/// </summary>
+[AttributeUsage(AttributeTargets.Enum)]
+[System.Diagnostics.Conditional("NETESCAPADES_ENUMGENERATORS_USAGES")]
+public sealed class EnumExtensionsAttribute : Attribute
 {
     /// <summary>
-    /// Add to enums to indicate that extension methods should be generated for the type
+    /// The namespace to generate the extension class.
+    /// If not provided the namespace of the enum will be used.
     /// </summary>
-    [System.AttributeUsage(System.AttributeTargets.Enum)]
-    [System.Diagnostics.Conditional("NETESCAPADES_ENUMGENERATORS_USAGES")]
-    public class EnumExtensionsAttribute : System.Attribute
-    {
-        /// <summary>
-        /// The namespace to generate the extension class.
-        /// If not provided the namespace of the enum will be used
-        /// </summary>
-        public string? ExtensionClassNamespace { get; set; }
+    public string? ExtensionClassNamespace { get; set; }
 
-        /// <summary>
-        /// The name to use for the extension class.
-        /// If not provided, the enum name with ""Extensions"" will be used.
-        /// For example for an Enum called StatusCodes, the default name
-        /// will be StatusCodesExtensions
-        /// </summary>
-        public string? ExtensionClassName { get; set; }
-    }
+    /// <summary>
+    /// The name to use for the extension class.
+    /// If not provided, the enum name with ""Extensions"" will be used.
+    /// For example for an Enum called StatusCodes, the default name
+    /// will be StatusCodesExtensions.
+    /// </summary>
+    public string? ExtensionClassName { get; set; }
 }
