@@ -2,7 +2,6 @@
 using System.Runtime.CompilerServices;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
-using Microsoft.CodeAnalysis.Emit;
 using NetEscapades.EnumGenerators;
 using NetEscapades.EnumGenerators.Benchmarks;
 
@@ -14,10 +13,8 @@ BenchmarkSwitcher
 public enum TestEnum
 {
     First = 0,
-
-    [Display(Name = "2nd")]
-    Second = 1,
-    Third = 2,
+    [Display(Name = "2nd")] Second = 1,
+    Third = 2
 }
 
 [MemoryDiagnoser]
@@ -84,89 +81,78 @@ public class IsDefinedNameBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public bool ExtensionsIsDefined()
     {
-        return TestEnumExtensions.IsDefined(_enum, allowMatchingMetadataAttribute: false);
+        return TestEnumExtensions.IsDefined(_enum, false);
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public bool EnumIsDefinedNameDisplayNameWithReflection()
     {
-        return EnumHelper<TestEnum>.TryParseByDisplayName(_enumDisplaName, ignoreCase: false, out _);
+        return EnumHelper<TestEnum>.TryParseByDisplayName(_enumDisplaName, false, out _);
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public bool ExtensionsIsDefinedNameDisplayName()
     {
-        return TestEnumExtensions.IsDefined(_enumDisplaName, allowMatchingMetadataAttribute: true);
+        return TestEnumExtensions.IsDefined(_enumDisplaName, true);
     }
 }
 
-#if NETCOREAPP && !NETCOREAPP2_0 && !NETCOREAPP1_1 && !NETCOREAPP1_0
 [MemoryDiagnoser]
 public class IsDefinedNameFromSpanBenchmark
 {
-    private static readonly char[] _enum = new char[] { 'S', 'e', 'c', 'o', 'n', 'd' };
-    private static readonly char[] _enumDisplayName = new char[] { '2', 'n', 'd' };
+    private static readonly char[] _enum = { 'S', 'e', 'c', 'o', 'n', 'd' };
+    private static readonly char[] _enumDisplayName = { '2', 'n', 'd' };
 
     [Benchmark(Baseline = true)]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public bool EnumIsDefined()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enum;
-        return Enum.IsDefined(typeof(TestEnum), _enumAsSpan.ToString());
+        ReadOnlySpan<char> enumAsSpan = _enum;
+        return Enum.IsDefined(typeof(TestEnum), enumAsSpan.ToString());
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public bool EnumIsDefinedNameDisplayNameWithReflection()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enumDisplayName;
-        return EnumHelper<TestEnum>.TryParseByDisplayName(_enumAsSpan.ToString(), ignoreCase:false, out _);
+        ReadOnlySpan<char> enumAsSpan = _enumDisplayName;
+        return EnumHelper<TestEnum>.TryParseByDisplayName(enumAsSpan.ToString(), false, out _);
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public bool ExtensionsIsDefined()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enum;
-        return TestEnumExtensions.IsDefined(_enumAsSpan.ToString(), allowMatchingMetadataAttribute: false);
+        ReadOnlySpan<char> enumAsSpan = _enum;
+        return TestEnumExtensions.IsDefined(enumAsSpan.ToString(), false);
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public bool ExtensionsIsDefinedSpan()
     {
-        return TestEnumExtensions.IsDefined(_enum.AsSpan(), allowMatchingMetadataAttribute: false);
+        return TestEnumExtensions.IsDefined(_enum.AsSpan(), false);
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public bool ExtensionsIsDefinedDisplayNameSpan()
     {
-        return TestEnumExtensions.IsDefined(_enumDisplayName.AsSpan(), allowMatchingMetadataAttribute: true);
+        return TestEnumExtensions.IsDefined(_enumDisplayName.AsSpan(), true);
     }
 }
-#endif
 
 [MemoryDiagnoser]
 public class GetValuesBenchmark
 {
-#if NETFRAMEWORK
-    [Benchmark(Baseline = true)]
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public TestEnum[] EnumGetValues()
-    {
-        return (TestEnum[])Enum.GetValues(typeof(TestEnum));
-    }
-#else
     [Benchmark(Baseline = true)]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum[] EnumGetValues()
     {
         return Enum.GetValues<TestEnum>();
     }
-#endif
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
@@ -179,21 +165,12 @@ public class GetValuesBenchmark
 [MemoryDiagnoser]
 public class GetNamesBenchmark
 {
-#if NETFRAMEWORK
-    [Benchmark(Baseline = true)]
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public string[] EnumGetNames()
-    {
-        return Enum.GetNames(typeof(TestEnum));
-    }
-#else
     [Benchmark(Baseline = true)]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public string[] EnumGetNames()
     {
         return Enum.GetNames<TestEnum>();
     }
-#endif
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
@@ -210,7 +187,7 @@ public class TryParseBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum EnumTryParse()
     {
-        return Enum.TryParse("Second", ignoreCase: false, out TestEnum result)
+        return Enum.TryParse("Second", false, out TestEnum result)
             ? result
             : default;
     }
@@ -219,7 +196,7 @@ public class TryParseBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParse()
     {
-        return TestEnumExtensions.TryParse("Second", out TestEnum result, ignoreCase: false)
+        return TestEnumExtensions.TryParse("Second", out var result, false)
             ? result
             : default;
     }
@@ -228,32 +205,31 @@ public class TryParseBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum EnumTryParseDisplayNameWithReflection()
     {
-        return EnumHelper<TestEnum>.TryParseByDisplayName("2nd", ignoreCase: false, out TestEnum result) ? result : default;
+        return EnumHelper<TestEnum>.TryParseByDisplayName("2nd", false, out var result) ? result : default;
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParseDisplayName()
     {
-        return TestEnumExtensions.TryParse("2nd", out TestEnum result, ignoreCase: false, allowMatchingMetadataAttribute: true)
+        return TestEnumExtensions.TryParse("2nd", out var result, false, true)
             ? result
             : default;
     }
 }
 
-#if NETCOREAPP && !NETCOREAPP2_0 && !NETCOREAPP1_1 && !NETCOREAPP1_0
 [MemoryDiagnoser]
 public class TryParseFromSpanBenchmark
 {
-    private static readonly char[] _enum = new char[] { 'S', 'e', 'c', 'o', 'n', 'd' };
-    private static readonly char[] _enumDisplayName = new char[] { '2', 'n', 'd' };
+    private static readonly char[] _enum = { 'S', 'e', 'c', 'o', 'n', 'd' };
+    private static readonly char[] _enumDisplayName = { '2', 'n', 'd' };
 
     [Benchmark(Baseline = true)]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum EnumTryParse()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enum;
-        return Enum.TryParse(_enumAsSpan.ToString(), ignoreCase: false, out TestEnum result)
+        ReadOnlySpan<char> enumAsSpan = _enum;
+        return Enum.TryParse(enumAsSpan.ToString(), false, out TestEnum result)
             ? result
             : default;
     }
@@ -262,8 +238,8 @@ public class TryParseFromSpanBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParse()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enum;
-        return TestEnumExtensions.TryParse(_enumAsSpan.ToString(), out TestEnum result, ignoreCase: false)
+        ReadOnlySpan<char> enumAsSpan = _enum;
+        return TestEnumExtensions.TryParse(enumAsSpan.ToString(), out var result, false)
             ? result
             : default;
     }
@@ -272,7 +248,7 @@ public class TryParseFromSpanBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParseSpan()
     {
-        return TestEnumExtensions.TryParse(_enum.AsSpan(), out TestEnum result, ignoreCase: false)
+        return TestEnumExtensions.TryParse(_enum.AsSpan(), out var result, false)
             ? result
             : default;
     }
@@ -281,16 +257,16 @@ public class TryParseFromSpanBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum EnumTryParseDisplayNameWithReflection()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enumDisplayName;
-        return EnumHelper<TestEnum>.TryParseByDisplayName(_enumAsSpan.ToString(), ignoreCase: false, out TestEnum result) ? result : default;
+        ReadOnlySpan<char> enumAsSpan = _enumDisplayName;
+        return EnumHelper<TestEnum>.TryParseByDisplayName(enumAsSpan.ToString(), ignoreCase: false, out var result) ? result : default;
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParseDisplayName()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enumDisplayName;
-        return TestEnumExtensions.TryParse(_enumAsSpan.ToString(), out TestEnum result, ignoreCase: false, allowMatchingMetadataAttribute: true)
+        ReadOnlySpan<char> enumAsSpan = _enumDisplayName;
+        return TestEnumExtensions.TryParse(enumAsSpan.ToString(), out var result, ignoreCase: false, allowMatchingMetadataAttribute: true)
             ? result
             : default;
     }
@@ -299,12 +275,11 @@ public class TryParseFromSpanBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParseDisplayNameSpan()
     {
-        return TestEnumExtensions.TryParse(_enumDisplayName.AsSpan(), out TestEnum result, ignoreCase: false, allowMatchingMetadataAttribute: true)
+        return TestEnumExtensions.TryParse(_enumDisplayName.AsSpan(), out var result, ignoreCase: false, allowMatchingMetadataAttribute: true)
             ? result
             : default;
     }
 }
-#endif
 
 [MemoryDiagnoser]
 public class TryParseIgnoreCaseBenchmark
@@ -313,7 +288,7 @@ public class TryParseIgnoreCaseBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum EnumTryParseIgnoreCase()
     {
-        return Enum.TryParse("second", ignoreCase: true, out TestEnum result)
+        return Enum.TryParse("second", true, out TestEnum result)
             ? result
             : default;
     }
@@ -322,7 +297,7 @@ public class TryParseIgnoreCaseBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParseIgnoreCase()
     {
-        return TestEnumExtensions.TryParse("second", out TestEnum result, ignoreCase: true)
+        return TestEnumExtensions.TryParse("second", out var result, true)
             ? result
             : default;
     }
@@ -331,31 +306,30 @@ public class TryParseIgnoreCaseBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum EnumTryParseIgnoreCaseDisplayNameWithReflection()
     {
-        return EnumHelper<TestEnum>.TryParseByDisplayName("2ND", ignoreCase: true, out TestEnum result) ? result : default;
+        return EnumHelper<TestEnum>.TryParseByDisplayName("2ND", true, out var result) ? result : default;
     }
 
     [Benchmark]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParseIgnoreCaseDisplayName()
     {
-        return TestEnumExtensions.TryParse("2ND", out TestEnum result, ignoreCase: true, allowMatchingMetadataAttribute: true)
+        return TestEnumExtensions.TryParse("2ND", out var result, true, true)
             ? result
             : default;
     }
 }
 
-#if NETCOREAPP && !NETCOREAPP2_0 && !NETCOREAPP1_1 && !NETCOREAPP1_0
 [MemoryDiagnoser]
 public class TryParseIgnoreCaseFromSpanBenchmark
 {
-    private static readonly char[] _enum = new char[] { 's', 'e', 'c', 'o', 'n', 'd' };
+    private static readonly char[] _enum = { 's', 'e', 'c', 'o', 'n', 'd' };
 
     [Benchmark(Baseline = true)]
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum EnumTryParseIgnoreCase()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enum;
-        return Enum.TryParse(_enumAsSpan.ToString(), ignoreCase: true, out TestEnum result)
+        ReadOnlySpan<char> enumAsSpan = _enum;
+        return Enum.TryParse(enumAsSpan.ToString(), true, out TestEnum result)
             ? result
             : default;
     }
@@ -364,8 +338,8 @@ public class TryParseIgnoreCaseFromSpanBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParseIgnoreCase()
     {
-        ReadOnlySpan<char> _enumAsSpan = _enum;
-        return TestEnumExtensions.TryParse(_enumAsSpan.ToString(), out TestEnum result, ignoreCase: true)
+        ReadOnlySpan<char> enumAsSpan = _enum;
+        return TestEnumExtensions.TryParse(enumAsSpan.ToString(), out var result, true)
             ? result
             : default;
     }
@@ -375,12 +349,11 @@ public class TryParseIgnoreCaseFromSpanBenchmark
     [MethodImpl(MethodImplOptions.NoInlining)]
     public TestEnum ExtensionsTryParseIgnoreCaseSpan()
     {
-        return TestEnumExtensions.TryParse(_enum.AsSpan(), out TestEnum result, ignoreCase: true)
+        return TestEnumExtensions.TryParse(_enum.AsSpan(), out var result, ignoreCase: true)
             ? result
             : default;
     }
 }
-#endif
 
 [MemoryDiagnoser]
 public class EnumLengthBenchmark
