@@ -1,4 +1,3 @@
-using System;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Reflection;
@@ -8,18 +7,16 @@ using FluentAssertions.Execution;
 namespace NetEscapades.EnumGenerators.IntegrationTests;
 
 #nullable enable
-public abstract class ExtensionTests<T> where T : struct
+public abstract class ExtensionTests<T> where T : struct, Enum
 {
     protected abstract string ToStringFast(T value);
     protected abstract bool IsDefined(T value);
     protected abstract bool IsDefined(string name, bool allowMatchingMetadataAttribute = false);
-#if READONLYSPAN
     protected abstract bool IsDefined(in ReadOnlySpan<char> name, bool allowMatchingMetadataAttribute);
-#endif
     protected abstract bool TryParse(string name, out T parsed, bool ignoreCase, bool allowMatchingMetadataAttribute);
-#if READONLYSPAN
     protected abstract bool TryParse(in ReadOnlySpan<char> name, out T parsed, bool ignoreCase, bool allowMatchingMetadataAttribute);
-#endif
+    protected abstract T? GetValueOrDefault(string name, bool ignoreCase, bool allowMatchingMetadataAttribute);
+    protected abstract T? GetValueOrDefault(in ReadOnlySpan<char> name, bool ignoreCase, bool allowMatchingMetadataAttribute);
 
     protected void GeneratesToStringFastTest(T value)
     {
@@ -27,8 +24,8 @@ public abstract class ExtensionTests<T> where T : struct
         var valueAsString = value.ToString();
 
         TryGetDisplayNameOrDescription(valueAsString, out var displayName);
-        var expectedValue = displayName is null ? valueAsString : displayName;
-        
+        var expectedValue = displayName ?? valueAsString;
+
         serialized.Should().Be(expectedValue);
     }
 
@@ -46,21 +43,19 @@ public abstract class ExtensionTests<T> where T : struct
         isDefined.Should().Be(expectedResult);
     }
 
-#if READONLYSPAN
     protected void GeneratesIsDefinedTest(in ReadOnlySpan<char> name, bool allowMatchingMetadataAttribute)
     {
         var isDefined = IsDefined(name, allowMatchingMetadataAttribute);
         var expectedResult = ValidateIsDefined(name.ToString(), allowMatchingMetadataAttribute);
         isDefined.Should().Be(expectedResult);
     }
-#endif
 
     private bool ValidateIsDefined(string name, bool allowMatchingMetadataAttribute)
     {
         bool expectedResult;
         if (allowMatchingMetadataAttribute)
         {
-            expectedResult = TryGetEnumByDisplayNameOrDescription(name, ignoreCase: false, out _);
+            expectedResult = TryGetEnumByDisplayNameOrDescription(name, false, out _);
             if (!expectedResult)
             {
                 expectedResult = Enum.IsDefined(typeof(T), name);
@@ -77,24 +72,22 @@ public abstract class ExtensionTests<T> where T : struct
     protected void GeneratesTryParseTest(string name, bool ignoreCase, bool allowMatchingMetadataAttribute)
     {
         var isValid = TryParse(name, out var result, ignoreCase, allowMatchingMetadataAttribute);
-        ValidateTryParse(name, ignoreCase, allowMatchingMetadataAttribute, out bool expectedValidity, out T expectedResult);
+        ValidateTryParse(name, ignoreCase, allowMatchingMetadataAttribute, out var expectedValidity, out var expectedResult);
 
         _ = new AssertionScope();
         isValid.Should().Be(expectedValidity);
         result.Should().Be(expectedResult);
     }
 
-#if READONLYSPAN
     protected void GeneratesTryParseTest(in ReadOnlySpan<char> name, bool ignoreCase, bool allowMatchingMetadataAttribute)
     {
         var isValid = TryParse(name, out var result, ignoreCase, allowMatchingMetadataAttribute);
-        ValidateTryParse(name.ToString(), ignoreCase, allowMatchingMetadataAttribute, out bool expectedValidity, out T expectedResult);
+        ValidateTryParse(name.ToString(), ignoreCase, allowMatchingMetadataAttribute, out var expectedValidity, out var expectedResult);
 
         _ = new AssertionScope();
         isValid.Should().Be(expectedValidity);
         result.Should().Be(expectedResult);
     }
-#endif
 
     private void ValidateTryParse(string name, bool ignoreCase, bool allowMatchingMetadataAttribute, out bool expectedValidity, out T expectedResult)
     {
@@ -110,6 +103,49 @@ public abstract class ExtensionTests<T> where T : struct
         {
             expectedValidity = Enum.TryParse(name, ignoreCase, out expectedResult);
         }
+    }
+
+    protected void GeneratesGetValueOrDefaultTest(string name, bool ignoreCase, bool allowMatchingMetadataAttribute)
+    {
+        var enumValue = GetValueOrDefault(name, ignoreCase, allowMatchingMetadataAttribute);
+        ValidateTryParse(name, ignoreCase, allowMatchingMetadataAttribute, out var expectedValidity, out var expectedResult);
+
+        _ = new AssertionScope();
+        if (enumValue is null)
+            expectedValidity.Should().BeFalse();
+        else
+        {
+            expectedValidity.Should().BeTrue();
+            enumValue.Should().Be(expectedResult);
+        }
+    }
+
+    protected void GeneratesGetValueOrDefaultTest(in ReadOnlySpan<char> name, bool ignoreCase, bool allowMatchingMetadataAttribute)
+    {
+        var enumValue = GetValueOrDefault(name, ignoreCase, allowMatchingMetadataAttribute);
+        ValidateTryParse(name.ToString(), ignoreCase, allowMatchingMetadataAttribute, out var expectedValidity, out var expectedResult);
+
+        _ = new AssertionScope();
+        if (enumValue is null)
+            expectedValidity.Should().BeFalse();
+        else
+        {
+            expectedValidity.Should().BeTrue();
+            enumValue.Should().Be(expectedResult);
+        }
+    }
+
+    protected void GeneratesGetMetadataNamesOrDefaultTest(string[] displayNames)
+    {
+        var expected = new List<string>(displayNames.Length);
+        foreach (var value in Enum.GetValues<T>())
+        {
+            var valueAsString = value.ToString();
+            TryGetDisplayNameOrDescription(valueAsString, out var displayName);
+            expected.Add(displayName ?? valueAsString);
+        }
+
+        displayNames.Should().Equal(expected);
     }
 
     protected void GeneratesGetValuesTest(T[] values)
@@ -132,7 +168,7 @@ public abstract class ExtensionTests<T> where T : struct
         var enumValues = (T[])Enum.GetValues(typeof(T));
         foreach (var value in enumValues)
         {
-            if (TryGetDisplayNameOrDescription(value.ToString(), out var displayName) && string.Equals(displayName,name, stringComparisonOptions))
+            if (TryGetDisplayNameOrDescription(value.ToString(), out var displayName) && string.Equals(displayName, name, stringComparisonOptions))
             {
                 enumValue = value;
                 return true;
@@ -142,37 +178,20 @@ public abstract class ExtensionTests<T> where T : struct
         return false;
     }
 
-    private bool TryGetDisplayNameOrDescription(
+    private static bool TryGetDisplayNameOrDescription(
         string? value,
-#if NETCOREAPP3_0_OR_GREATER
-        [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out string? displayName)
-#else
+        [System.Diagnostics.CodeAnalysis.NotNullWhen(true)]
         out string? displayName)
-#endif
     {
         displayName = default;
 
-        if (typeof(T).IsEnum)
-        {
-            // Prevent: Warning CS8604  Possible null reference argument for parameter 'name' in 'MemberInfo[] Type.GetMember(string name)'
-            if (value is not null)
-            {
-                var memberInfo = typeof(T).GetMember(value);
-                if (memberInfo.Length > 0)
-                {
-                    // Doesn't take order into account, but we don't test with both currently
-                    displayName = memberInfo[0].GetCustomAttribute<DisplayAttribute>()?.GetName()
-                        ?? memberInfo[0].GetCustomAttribute<DescriptionAttribute>()?.Description;
-                    if (displayName is null)
-                    {
-                        return false;
-                    }
-                    
-                    return true;
-                }
-            }
-        }
-
-        return false;
+        if (value is null) return false;
+        // Prevent: Warning CS8604  Possible null reference argument for parameter 'name' in 'MemberInfo[] Type.GetMember(string name)'
+        var memberInfo = typeof(T).GetMember(value);
+        if (memberInfo.Length <= 0) return false;
+        // Doesn't take order into account, but we don't test with both currently
+        displayName = memberInfo[0].GetCustomAttribute<DisplayAttribute>()?.GetName()
+                      ?? memberInfo[0].GetCustomAttribute<DescriptionAttribute>()?.Description;
+        return displayName is not null;
     }
 }
