@@ -1,25 +1,23 @@
-using System;
 using System.Collections.Immutable;
-using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 
 namespace NetEscapades.EnumGenerators.Tests;
 
-internal class TestHelpers
+internal static class TestHelpers
 {
     public static (ImmutableArray<Diagnostic> Diagnostics, string Output) GetGeneratedOutput<T>(string source)
         where T : IIncrementalGenerator, new()
     {
         var syntaxTree = CSharpSyntaxTree.ParseText(source);
         var references = AppDomain.CurrentDomain.GetAssemblies()
-            .Where(_ => !_.IsDynamic && !string.IsNullOrWhiteSpace(_.Location))
-            .Select(_ => MetadataReference.CreateFromFile(_.Location))
+            .Where(static _ => !_.IsDynamic && !string.IsNullOrWhiteSpace(_.Location))
+            .Select(static _ => MetadataReference.CreateFromFile(_.Location))
             .Concat(new[]
             {
                 MetadataReference.CreateFromFile(typeof(T).Assembly.Location),
                 MetadataReference.CreateFromFile(typeof(EnumExtensionsAttribute).Assembly.Location),
-                MetadataReference.CreateFromFile(typeof(System.ComponentModel.DataAnnotations.DisplayAttribute).Assembly.Location),
+                MetadataReference.CreateFromFile(typeof(System.ComponentModel.DataAnnotations.DisplayAttribute).Assembly.Location)
             });
 
         var compilation = CSharpCompilation.Create(
@@ -36,6 +34,6 @@ internal class TestHelpers
 
         var trees = outputCompilation.SyntaxTrees.ToList();
 
-        return (diagnostics, trees.Count != originalTreeCount ? trees[trees.Count - 1].ToString() : string.Empty);
+        return (diagnostics, trees.Count != originalTreeCount ? trees[^1].ToString() : string.Empty);
     }
 }
